# Performance improvement for img_info