Better use of numpy vectorization

```
In [21]: import numpy as np

In [22]: a = np.full((640, 480), 150)

In [23]: %timeit np.all(np.vectorize(lambda x: x == a[0][0])(a))
102 ms ± 1.02 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [24]: %timeit np.all(a == a[0][0])
90.6 µs ± 499 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```
So this is a speedup by several orders of magnitude (100X). I'm sure this is not necessary a bottleneck, but I thought you might like the change.